todo de servicios

+++ app/Http/Controllers/ServicesController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 use App\Models\Service;
+use Illuminate\Support\Facades\Validator;
 
 use Illuminate\Http\Request;
 
@@ -9,11 +10,17 @@ class ServicesController extends Controller
 {
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => 'required|string|max:255',
+        $validator = Validator::make($request->all(), [
+            'name' => 'required|string|max:50',
             'description' => 'required|string',
             'price' => 'required|numeric|min:0',
         ]);
+    
+        if ($validator->fails()) {
+            return response()->json([
+                'error' => $validator->errors()
+            ], 400);
+        }
 
         $service = Service::create([
             'name' => $request->name,
@@ -24,38 +31,56 @@ class ServicesController extends Controller
         return response()->json($service);
     }
 
-    public function update(Request $request, $id)
+    public function update(Request $request)
     {
-        $request->validate([
-            'name' => 'string|max:255',
-            'description' => 'string',
-            'price' => 'numeric|min:0',
+        $validator = Validator::make($request->all(), [
+            'name' => 'required|string|max:50',
+            'description' => 'required|string',
+            'price' => 'required|numeric|min:0',
         ]);
+    
+        if ($validator->fails()) {
+            return response()->json([
+                'error' => $validator->errors()
+            ], 400);
+        }
 
-        $service = Service::findOrFail($id);
-        $service->update($request->all());
+        $service = Service::find($request->id);
+
+        if(!$service){
+            return response()->json(['mensaje' => 'Servicio no encontrado'], 400);
+        }
+
+        $service->update($validator->validated());
 
         return response()->json($service);
     }
 
     public function destroy($id)
     {
-        $service = Service::findOrFail($id);
+        $service = Service::find($id);
+
+        if(!$service){
+            return response()->json(['status' => false], 400);
+        }
         
         $service->delete();
 
-        return response()->json(null, 204);
+        return response()->json(['status' => true], 204);
     }
 
     public function index()
     {
         $services = Service::all();
-        return response()->json($services);
+        return response()->json(['services' => $services, 'status' => true] );
     }
 
     public function show($id)
     {
-        $service = Service::findOrFail($id);
+        $service = Service::find($id);
+        if(!$service){
+            return response()->json(['mensaje' => 'Servicio no encontrado'], 400);
+        }
         return response()->json($service);
     }
 }
